UPSTREAM: ARM: mvebu: Clean-up the Armada XP support
This patch removes the unneeded include of the armada-370-xp.h header.

It also moves some declarations from this file into more accurate
places.

Finally, it also adds a comment explaining that we can't remove yet the
smp field in the dt machine struct due to backward compatibly of the
device tree.

In a few releases, when the old device tree will be obsolete, we will be
able to remove the smp field and then the armada-370-xp.h header.

Signed-off-by: Gregory CLEMENT <gregory.clement@free-electrons.com>
Tested-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Reviewed-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Link: https://lkml.kernel.org/r/1414669184-16785-2-git-send-email-gregory.clement@free-electrons.com
Signed-off-by: Jason Cooper <jason@lakedaemon.net>
Signed-off-by: Nadav Haklai <nadavh@marvell.com>

BUG=chrome-os-partner:43662
TEST=build and boot Armada385-DB-GP board

Change-Id: I3bce64a4a6127a24d1590204d6f12d970933e548
Reviewed-on: https://chromium-review.googlesource.com/289528
Commit-Ready: Nadav Haklai <nadavh@marvell.com>
Tested-by: Kan Yan <kyan@chromium.org>
Reviewed-by: Kan Yan <kyan@chromium.org>
Reviewed-by: Bryan Freed <bfreed@google.com>