CHROMIUM: devfreq: Add active watermark devfreq governor
This patch adds a new watermark governor that actively alters the
watermark values based on the current frequency and target load.

The governor takes target load as a given property (80% by default)
and every time when the governor re-estimation is triggered, the
governor checks the current load and calculates the next frequency
that would push the device as close this target load as possible.

At this point also the watermark values are updated so that next
time the interrupt should come when the load change is significant
enough to cause frequency change to either next or previous
frequency.

BUG=chrome-os-partner:42645
TEST=Build and boot on smaug.

Change-Id: If324544ba747358b792486fa4b9bcbaf21d130e0
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/305366
Commit-Ready: Andrew Chew <achew@nvidia.com>
Tested-by: Andrew Chew <achew@nvidia.com>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>