UPSTREAM: bus: mvebu-mbus: provide a mechanism to save SDRAM window configuration
On Marvell EBU platforms, when doing suspend/resume, the SDRAM window
configuration must be saved on suspend, and restored on
resume. However, it needs to be restored on resume *before*
re-entering the kernel, because the SDRAM window configuration defines
the layout of the memory. For this reason, it cannot simply be done in
the ->suspend() and ->resume() hooks of the mvebu-mbus driver.

Instead, it needs to be restored by the bootloader "boot info"
mechanism used when resuming. This mechanism allows the kernel to
define a list of (address, value) pairs when suspending, that the
bootloader will restore on resume before jumping back into the kernel.

This commit therefore adds a new function to the mvebu-mbus driver,
called mvebu_mbus_save_cpu_target(), which will be called by the
platform code to make the mvebu-mbus driver save the SDRAM window
configuration in a way that can be understood by the bootloader "boot
info" mechanism.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Reviewed-by: Gregory CLEMENT <gregory.clement@free-electrons.com>
Link: https://lkml.kernel.org/r/1416585613-2113-8-git-send-email-thomas.petazzoni@free-electrons.com
Signed-off-by: Jason Cooper <jason@lakedaemon.net>
(cherry picked from commit 4749c02b8da6d8dbc29218652985bda844017e95)
Signed-off-by: Nadav Haklai <nadavh@marvell.com>

BUG=chrome-os-partner:43662
TEST=Test suspend to RAM on Armada385-DB-GP board

Change-Id: Ic5877ffe6f264aae16faa83a4182758ea3a0228a
Reviewed-on: https://chromium-review.googlesource.com/286522
Commit-Ready: Nadav Haklai <nadavh@marvell.com>
Tested-by: Kan Yan <kyan@chromium.org>
Reviewed-by: Kan Yan <kyan@chromium.org>
Reviewed-by: Bryan Freed <bfreed@google.com>
Reviewed-by: Grant Grundler <grundler@chromium.org>